requirements.txt
google-cloud-bigquery
pandas
db-dtypes
pyodbc
dbt-sqlserver
python-dotenv
sqlalchemy